+++ src/Utility.ts
@@ -1,0 +1,13 @@
+import sharp, { Sharp } from "sharp";
+import ansiAlign from "ansi-align";
+import wrapAnsi from "wrap-ansi";
+
+export function isSharpInstance(instance: object): instance is Sharp {
+  return instance instanceof sharp;
+}
+
+/** Centres and wraps a piece of text according to width */
+export function centreText(text: string, width: number): string {
+  const wrappedLine: string = ansiAlign.center(wrapAnsi(text, width) + "\n" + " ".repeat(width));
+  return wrappedLine.substring(0, wrappedLine.lastIndexOf("\n"));
+}